+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - [#10] fix(modrinth): assume mods need to be loaded on both sides (see comments in code for reasoning)
 - [#12] fix(installer): removing a file from the pack now correctly removes it when updating with `sculk install` (files not managed by Sculk will not be touched).
+- [#11] fix(refresh): add `--watch` argument to watch the file system for changes and refresh when they are detected.